Replace nugram.flxh with separate grammar files.
We now allow --import=@response-file which means to consider
the response file as a list of files to be imported. Each line
of the response file is either a file to be imported, or, if it
looks like @another-response-file, then recursively another
response file. You can only use that syntax in a response file,
not an actual felix file.

Splitting the grammar up makes it easier to document.