Initial implementation of sarray.
Also some more mods to the compiler to support operator dot and
addressing structure components.

sarray is an ultra high performance sparse array, implemented with
a darray to contain the data and JudyL array to map input index to
slot number in the darray. In case elements are deleted, a Judy1
array keeps track of free slots so they can be reused. A pack method
rebuilds the array with minimum memory and in index order.

We currently provide only get, set, and del methods.
The only constructor at the moment just accepts the default value.
Note that del'eting an index is not the same as setting it to a value
that happens to be the default. This is because we do not assume
the existence of an equality method.