Add 'requires scanner "C_func_name"' to C type binding.
This allows you to specify the scanner to be used on a type.
At present you can't specify any client data, and there's
no particular way to handle polymorphic types.

The "C_func_name" can actually contain any of the usual micro-encodings
for type variables, but they're ignored.

We have added a scanner for Judy1 arrays. More is needed though.
For example, the finaliser should clear the array, but there's no
way to specify that. Adding new properties is a serious pain,
they get translated and shifted about, and printed in each form.

Also, there are two types of Judy1 arrays: those using integer keys
(which don't require scanning), and those using pointers (which do,
at least if they're Felix pointers). It is conservatively safe
to scan integers (may cause a leak at worst), but it's expensive
so a better way needs to be found. One possibility is to turn
Judy1 into a typeclass and then specialise on the two legal
cases (integers and pointers). Actually packed variants would
work too (but that's pretty implementation dependent).