Some changed to typeclass organisation.
Arrays and tuples up to length 5 now have equality and str.
Arrays have comparisons. This some causes the axiom check
examples to fail. I can't track it down at the moment.

Axiom check is a bit hacked in the compiler. It generates
ordinary code, and then we can get a "failure to find an instance".
Although the axiom_check generator actually checks there is an
instance, it probably only does this check first order.
If an instance function requires *another* instance it doesn't
discover that, and then we get a failure down the track.

Somehow this shouldn't happen: either the check should be recursive,
which is a serious pain, or we should just throw the check case out.
The problem is the fact it was just an axiom check is lost by the time
the failure is detected.

What I really don't understand is which axiom is causing the problem:
the instance ts=int^2->int will never have equality defined: its a
function! Well, the check in the example actually uses a function
for equality, but there's no lookup for equality OF functions.