Add procedures to spawn processes.
This is the start of a nasty chain of events which will lead to
rather too much rewriting of the flx harness to work around the
quoting problem which occurs when flx is used to run a command
directly, compared to compiling a program which is run separately.
For example flx_cp has arguments which are Perl regexps and need
nasty quoting already, but because flx calls it using a shell string,
the result is different to calling it directly, due to an extra level
of shell. This inconsistency can't be tolerated.

Unfortunately flx uses string concatenation to build commands,
so there's no easy solution. The "right way" is to either add
extra quoting to fool the shell, or, to pass arguments in string
arrays.

Unfortunately Posix has lousy support for this. We have to encode
exec functions to replace the existing process, fork so the main
process isn't clobbered, and wait, so that the main process can
wait for the child. That's what this patch is starting to do.

Again, it is VERY unfortunate that flx is using strings, because
commands like "g++ -g" simply won't work with the exec family.
We're going to have to fix the fbuld configuration to separate
all the arguments, and keep them separated.

Even more unfortunate .. it isn't clear exec/fork etc will work
on Windows, whereas system() command will (because it is standard C,
whereas exec/fork etc are Posix only). Windows coding surely exists
but I don't have a Windows box so I can't even begin to think about
how to provide a platform independent abstraction here.