Add semantics for C expressions.
This is the easy case so I'll do it first. Some mods to compiler may
be needed for C operations not currently supported in Felix,
or not supported in a compliant way.

Parsing expressions has two immediate uses: for array bounds,
which is type information, and for emitting C functions directly.

The current implementation translates to Felix, it doesn't preserve
the C code as written. That could also be done but has a quite
different impact. In that model the Felix backend C expression
encoding should be generated.