Add for loops that capture variables.
That is, the control variable goes out of scope at the end of the loop.
The loop body is a block, so you can't jump into it. Any symbols,
including variables, declared in the loop are lost at the end of the
loop. This is similar to a C/C++ loop.