Fix bug with typeclass virtual instantiation in some cases.
Instantiation of typeclass virtuals is done by unification.
However unification cannot handle extended types such as
type matches and type functions. We have attempted to
define complex[t in floats] by a type match but binding
a call to an instance fcomplex, float to Complex[complex[t],t]
failed. We do a trick in the instantiator where we split out
assignments, substitute into other terms of the remaining
equations, and then reduce their left and right hand terms,
removing type matches and applications of type functions
for which the substitutions remove the type variables.

The trick is incomplete: it should be done inside the unification
algorithm, however that is a bit more work, and it tricky! It requires
introducing the beta-reduction code into the unification algorithm,
which is not so hard. What IS hard is ensuring we don't get an
infinite loop. At present every step either: destroys one equation
where the top level structures match, and introduces a set of
assignments where type variables correspond to terms,
or, it removes an assignment in one of two ways: either the
variable is dependent in which case it is substituted away
in the other equations and the assignment discarded, or
it is an independent variable and after the substitution it
is added to the MGU.

This algorithm clearly terminates since terms are finite and
so there is a fixed limit to the maximum number of steps to
reduce any pair of terms to assignments (or a mismatch),
and assigments are removed each iteration, so the maximum
number of iterations can be fixed entirely by the maximum
term depth and number of equations.

Type match and lambda terms, on the other hand, cannot be
discarded simply because they neither match nor can
be reduced, since they might become reducible after a
later substitution. however they can't just be thrown back
into the set of equations in that case because they'd
just be reconsidered later and if they remain the same
get thrown back again: an infinite loop.

The "right" solution is probably to add them to a "bad" set
and not reconsider them unless they contain a variable
which is substituted away. We note that normally such substitutions
can't introduce new variables (only eliminate them) and so this
should be safe (no infinite stream of new variables).

Bottom line: the trick implemented is enough to make complex[t]
work in at least one case whereas previously it did not.