Refactor array class. Add bounded sparse array (bsarray).
Standard sparse array can't support arbitrary iterations (iter, fold,
etc). However it CAN support iteration over "nondefault" elements in
some sense. For fold_left the requirement would be that f acc dflt =
acc, for example addition of 0: this allows the default values to
be skipped.  At this stage we don't know how to enforce this although
we might express it with an axiom (except that we'd need a function
to extract the default).

Perhaps better than that: a requirement the function is a member
of the typeclass with that constraint.

It's a bit ugly that typeclass and instance are distinguished:
polymorphic instances are subclasses and it should be possible
to add new features (eg new virtuals) and subclass it again.