Change completely addressing of struct components.
If px is a pointer to a struct with field a, then px.a is a pointer
to the a component. This allows us to write px . a <- 1. We hope this
will go over to x . a = 1 translating to (&x) . a <- 1. I don't see
how this could happen at the moment. The idea is that & can ONLY be
applied to a whole variable (or possibly an application of a C primitive
with lvalue property).