Put q strings back in with constraint.
Perl interpolation is back, but with the restriction that
only simple variable names can be used. This is because
anything else would require parsing, and that's a bit of
a pain to get working. So this now works:

val x = 1;
val y = "World";
val z = 99.76;
val s = q"ab$(x)$(y)$(z)cd";
println$ s;