Add support for unions to be mapped to null pointers.
For exactly two cases, the first of which is a constant
constructor, the null pointer representation is used.
This applies to standard library list and opt types.
It opens the way to allow &T to be consider a never null
object pointer and a type null_or_object_pointer[T] intoduced
to support pointers, including those from C, to be null.
Then the type system we have now can deal with null pointer checking
without modification, and it will remain to provide some
convenient syntax.