Fix addressing rules in preparation for enforcement.
a) We allow x.j where x is a struct, cstruct, record or tuple,
and j is a field name (integer for tuple case).

b) We allow px.j in all these cases as well, where px
is a pointer to struct, cstruct, record or tuple,
the result is a pointer to the component.

c) We define &(x.j) = ((&x).j)

d) This allows assignments like x.a = 1 to work

e) As a NASTY side effect: &(x.f) = f (&x).

This should be STOPPED somehow because, unfortunately,
addressing an expression which is the application of
a C function marked "lvalue" is permitted.