Allow functions to be written inside structs.
The function is exported into the enclosing space
with the addition of an argument of the struct type
named and named "self".

In this first cut, polymorphism isn't supported.
In this first cut, procedures are not supported,
more precisely they accept a value like functions,
whereas for a procedure "self" should be a pointer
to allow mutation. It's not clear what to do with
generators.