Eliminate hard coded lexing of integer literals.
Special case integer literals not fixed yet,
these include case indices (sinteger).

First step in lifting most lexing from the hard
coded Dypgen grammar to the dynamically loaded one.
We cheated a lot here, and equiped the sex2flx routine
with the decoder, instead of writing it in Scheme.
This means the lexicology is still fixed, but the actual
lexing is lifted.