Ooops.. revert last change.
To reuse the executable, its time stamp must be newer (greater)
than the cache time. If the cache is cleared or rebuilt, it's
probably because the compiler or something got upgraded so the
executable should be rebuilt.

The test is NOT exhaustively correct, but it should allow developers
to ensure stale executables aren't used when running tests.