Lift identifiers to user space.
This pretty much finishes off the lifting.

We are forced to leave the full complexity of string kinds
in the hard coded parser because it's used extensively
in the grammar bootstrap.

Bootstrap identifiers are restricted to a more basic kind.
It remains to check everything works and then try to enforce
the independence.