Get rid of all the front end pretty printer stuff.
It is not used anywhere. Erick put this in with the idea
of replacing the old "print" code, but the job was
never finished. Pretty printers in Ocaml and straight
output don't mix, the buffering leads to jumbled messages.
I personally hate the Ocaml formatter stuff (meaning
I don't understand it :)

The code is removed because it increases the amount of
work required to modify the front end, with no useful
results. I'll probably remove the back end use as well.