Partial solution to the inherit problem.
Doesn't infinite loop now, but also doesn't actually
yank in the stuff it should. Not sure why, no error
until lookup depends on it, and then we just get a not
found error, not an error constructing the environment
in which to do the lookup (which is what I would expect).

Probably this is to do with hacks in which failures propagate
to some level and get handled a particular way (no matter what
kind of failure it is).