Add -o filename option to flx.
This option is useful for creating accessible
executables. Without this option either

(a) the executable lives next to the *.flx file OR
(b) if --output_dir is specified, the absolute
    filename as in (a) is nested inside the output_dir.

(a) can be precluded if the parent directory is write
protected. (b) solves this but it hides the file
in the cache with a long winded filename, since the
extension to an absolute filename is done first,
before nesting "as if" the output_dir were root  /.

The -o option therefore allows the executables to
persist over cache rebuilds without requiring the
source parent directory to be writable.

Note that if the filename is a path the path
components have to exist. Note also the filename
is used "as written" as therefore should be differ
between static and dynamic linkage.