Fix install so webserver works.
We put the plugins in /usr/local/lib on the assumption
the loader path will include that. The build system
is adjusted to make the plugins now. Note also it makes
build/release/shlib for the plugins. However the install
actually sources the binaries for the webserver from
build/release/tools. YMMV :)