Change grammar specification language all strings raw.
For grammar specifications, all the strings are now
raw, that is, there is no escape processing.
So a \ is just another character. This ONLY applies
to the grammar/lexical specifications. This is done
to simplify the hard coded grammar by removing
the need for escape processing, simplifying
the hard lexical structure significantly.