Add a hack to Flx_filesys.join.
If RHS is Windows filename (not Unix as it's supposed to be),
the translate C: into C\ so that we can append an absolute
native filename to the cache directory name without getting
an embedded C: in there. Actually it isn't clear why that
would be wrong... Windows allows anything in pathnames,
so why not C:\ ?? So maybe this hack won't work.