Rewrite the compilers symbol garbage collector.
The old one allows a couple of unused variables to
remain in the thread frame object. Apart from looking
ugly, these two variables (a) instantiated the asynchronous
I/O de-multiplexer and (b) started up pool of threads ready
to accept jobs. This is really bad with plugins because
each plugin did that as well, and for every instance!

The new collector is slower, because it iterates to a fixpoint,
ensuring all unused variables are removed. This is tricky,
because each time we find an unused variable and remove it
and its initialisation, another variable, previously used
in the initialisation, may become an orphan if it wasn't
used anywhere else. Then its initialiser needs to be deleted
too, and so on.

The new collector does not preserve all instances of
virtual functions, it only preserves all instances of
functions for which the base virtual is actually used,
or which are called directly.

Eventually all the virtuals must disappear, and uses
will be replaced by direct uses of the instances.

The aggressive stripping of variables in multiple places
is probably overkill (expensive in time for little gain),
it really only needs to be done once just before code
generation.

Also stripped out more unused routines.