Fix the fpc crashing bug.
Duh .. the config file specified to call xlat_fpc
in library fdoc2html.. so we should have got an exception
thrown (not a crash). Hard to tell at the moment why this
didn't appear to happen (it may have happened, and it's just
crappy error reporting translating to an abort() or something).