Add incrementable pointer type. Add transpose function to array.
This is a massive system wide invasive change to pointer
semantics. The type &T is no longer able to support
arithmetic. Instead you must use the type +T.

The most common impact is that you need interfaces
to use +char instead of &char. The latter is useful
only for pointing at a single character. The type +T
is a pointer to an array element. It is also called
carray[T].

The function prefix + can be applied to a &T to
convert it to a +T, this is a nasty cast "throwing
away" protection from pointer arithmetic. It should
not be necessary and may be removed in the future
however right now it may help upgrading code broken
by these changes.

The function prefix - can be applied to a +T to
get back a &T. This cast is unsafe unless you're sure
the pointer is actually poining at an object.