Generalise get_n term to handle general index expressions.
Modify the compiler to support indexing arrays with
unitsums. Such indexing is safe and does not require a
bounds check. The generalisation of get_n will subsequently
be used to allow indexing multi-dimensional arrays with
tuple arguments.

The array library will have to be changed to work by coercing
integer arguments to the correct sum type. The coercion has to
be checked. We also need a coercion representing "shape changing"
arrays, so that for example an array of arrays can be turned
into a matrix. Some more study of the isomorphisms involved is
required.