Begin solidifying the handling of what i am calling linear types.
A linear type (in my terminology) is a sum or product of units,
recursively, which can be represented compactly as an integer using
a type based isomorphism based on the component sizes, and thus
used to index generalised arrays, which are arrays indexed by
a linear type .. linearly. This allows iteration to use a simple
loop, so folds, maps, etc become universal. in other words,
it gives us polyadic arrays: ways to write code that works
with arrays of arbitrary, unknown linear indexes: in particular
matrices of arbitrary rank.

The encoding is simple enough (using only addition, multiplication,
division and remainder on integers) that prospects to move to
dependent typing (bounds known only at run time) are good
provided the type checker is equipped, down the track, with
suitable theorem proving capabilities to identify or relate
such types parametrically (i.e. to prove two matrices have
the same index type, without knowing the actual dimensions).
There's also the usual issue of "where to put the  variables
containing the shape information" eg for an array,
how do we find the length? How do we tie the lengths of
two arrays together (required to do parallel iteration,
for example to zip them into an array of pairs).

The current Felix code only handles base structural types:
Felix tuples and their dual cases. However this can and should be extended
to handle records and their duals variants, and structs and
their dual unions.