Drop all support for both darwin and clang.
The only option now is posix with gcc,
or windows. This is temporary. The problem is fbuild
gets it all wrong, it tries to determine the compiler
and flags to use based on the OS which is just wrong.
Fixing this in fbuild isn't trivial.

I will try to re-enable clang for Linux, after I get
a build working for OSX. I need that because several
patches have gone in and I have a serious bug to fix
and these patches destablised the whole process.

The bug is an infinite loop in some simple code
that appears to be a problem with the optimiser
"losing" code, I need to sort that out highest priority
because it undermines the whole project if its happening.
The bug appears to have been triggered by allowing code
that previous wouldn't build using the new compact linear
types to build, by changing array lookup in the Felix standard
library to use compiler generated code instead of binding to
C array lookup (which can't work for compact linear types).

However, whilst the new lookup code is probably wrong,
the generator loop driving the test case has got messed
up and these should be entirely unrelated.

Sorry but i have to investigate this first.

It may be this doesn't solve my problem: at the moment
I can build flx and flx-pkgconfig but the test case
in the build system just hangs forever, and hand builds
of other code wouldn't link.