Remove the faulty axioms and derivatives.
Add back later. Some of the derivativees as defined
in Wikipedia appear to make sense only for Real.
Wikipedia article is unclear if it would work for,
say, complex. So just keep these separate.

The axioms require approximate equality. To  be done
later, so for now just get rid of them.