Force -fno-strict-aliasing.
Felix breaks ISO C/C++ strict aliasing rules.
With optimisation, g++ actually uses this rule, and
this causes a bus error on Linux with gcc 4.2, so we
have to turn optimisations based on this rule off.

For C++9x there is no workaround due to faults in the
C++ language. For C++11 we can use cheat with unions.
Technically this still breaks strict aliasing rules but
probably won't be recognized by gcc.

Felix uses aliasing of structures which are layout
compatible. In particular tuples are structs with
a sequence of individual members and a constructor,
whereas arrays are arrays wrapped in a struct.
In Felix these are the same type, and we use a cast
to convert between them (breaking strict aliasing rules).