Fix really nasty bug in optimiser re Felix generators.
The optimiser treated generators written in Felix (not C bindings)
as if they were ordinary functions. In particular it failed
to enure generator calls were replaced by a variable previously
assigned the generator value. For some reason that case was
commented out of the special handling for generators.
Consequently this program:

var cnt = 1;
gen seq() = { ++cnt; return cnt; }
println$ #seq, #seq;

printed 3,3 instead of 2,3. The mutable code was
correctly duplicated, but the result of the call
wasn't assigned to a separate variable as it should
have been. If the "gen" above is changed to a fun,
3,3 makes sense. In that care the desired behaviour
of 2,3 can be recovered by

println$ (var #seq), (var #seq);

where "var" forces eager evaluation of the following
subexpression. This is essential at present if the
generator is wrapped in a closure because there is
no type distinguishing generators from functions:
the special handling only works for direct calls,
not calls through variables (closures).