Add line numbers to felix code.
The purpose is to allow HTML anchors to allow hyperlinking to
symbol definitions from the library index. The numbers are
displayed at the moment, which interferes with cut and paste.
It should be possible to turn them off, eg by clicking in
the number column, however then we'd need a way to turn them on.