Refactor fdoc2html, pull scanner out.
The result is slower than the previous state machine
model but is easier to work with. Also remove @link
command, since it doesn't fit into the scheme of things.
Use html <a> links instead. Fdoc commands are only suitable
for construction of vertical box lists <div>, not
horizontal <span> or inline markup.