Add new statement: with stmts-1 do stmts-2 done
Symbols introduced in stmts-1 are only used to define
stmts-2. For example here:

var tmp = 1;
var y = tmp;

we have a problem if we want to define y, but don't want
tmp to be visible. This:

with var tmp = 1 do
  var y = tmp;
done

solves it. The implementation is:

class dummy { stmts-1' stmts-2 } inherit dummy;

where stmts-1' are stmts-1 made private do the inherit
doesn't export them. Note any privates in stmts-2 will
not be exported either as an artefact of the implementation.
So actually this is equivalent:

with ; do private var tmp = 1; var y = tmp; done