Change the layout of fdocs.
I'm no artist! So there's a long way to go stylistically,
however we now get consistent navigation buttons with
mouseovers, in fixed (non-scrolling) positions in the browser.
Changed the font to sans-serif, background to a light grey,
text to dark grey.

Done with yet another plugin with the java-like object
interface stuff. Modification to the buttons only
takes a few seconds to build and install.