Make fdoc2html return an object.
We want flx_libcontents et all to be able to format documentation
comments in Felx programs "as fdoc" with all its features.
Well, at least to try it :)

To do this we need to return the core HTML without the
navigation bars etc.

It will be interesting, since fdoc can contain felix code,
which can have embedded comments, which are themselves
fdoc which can contain felix code .. :)

Anyhow, we just want the library docs to contain slightly
nicer formatted documentation, currently we just dump
a preformatted image of the comment lines (minnus
the leading //$).