Refactor inlining check so the same test is used everywhere.
Add diagnostic aborting program if a function marked
"inline" cannot be inlined. Work is incomplete because
it doesn't yet check for a function containing yield
(that check is done elsewhere and if true the inline_check
function never gets called so any conflict will not be detected).
To be fixed.

In the process I found a bug in the calculation used in
all but one place which has now been corrected, and may
have accounted for the infinite expansion of the inliner
in some case (since that was the symptom when I used the
wrong calculation everywhere after refactoring).