Compiler mods to enforce inline semantics.
Incomplete and not yet entirely forceful,
but the intent is to allow inline functions to
do stuff that might work if the function is inlined
into a top level procedure, such as issue service calls.

This is driven by the need for a function like
future.get() to trigger an schannel read.
In fact, read schannel is itself a perfect example!