Rework the whole system to support uniquely name entry points.
The aim is to allow static linkage of plugins.
We can also support storing multiple "programs" in a
single static archive too.

This change introduces two dynlink_t::link functions,
one in which you specify the modulename, and one
in which it is calculated from the filename.
The modulename is used as a prefix for the entry
points modulename_create_thread_frame for example.

This update does NOT fully upgrade plugin support,
it's just a pre-requisite for it.