Remove reference manual generation from user build process.
Instead, they're built by developer and put in repo
so the user build can just copy them. The build tools
used are the ones in the build/release but they use
shared libs and plugins in the (DY)LD_LIBRARY_PATH,
which are not up to date. To get the latest and greatest,
you have to build and install and then generate the docs
and install again.