Wrap tests.
Wrap tests in fdocs and put them in src.
Run make gentest to extract the flx and expect files.
Commit that too.

Idea is that we can install the test sources as fdocs
so the webserver can format them. Could work the same
as check-tut, but fbuild test module does better summaries.

Uncomfortable committing BOTH the fdoc source and the
products.