flx can now check output of a prog matches an expect file.
Use --stdout=foo --expect=bar to compare output saved in foo
with expected results in bar. If --stdout is left out it is
put in the cache (next to the C++ file if that's cached).

If you put --expect without a filename, the expect file
is taken to be the program filename with the extension
replaced by .expect.

This feature will be more useful when I add the ability
to specify a regex instead of just a single filename.
Then we can verfiy a whole set of regression tests
with a single command. Or compile a whole set of
programs with a single command.