Move the functionality of flx_pkgconfg into the library.
To follow, remove callouts to it in flx, just call the
library instead. Note in the build process we still have
to build flx_pkgconfig executable so Python can
build flx executable using it. Once flx is built
it will only be needed as a stand alone utility.

The call format is trivial:

var return_code, result = FlxPkgConfig::flx_pkgconfig
(tail #System::args);

i.e. just give a list of strings which would be the
arguments of the executable. The resulting return
code is the same as the executable would return,
and the resulting list of strings is the same
as the standard output would have been,
separated by spaces.