Add flx --staticlib option.
This option creates a static link library (.a on Unix,
or .lib on Windows) instead of an executable.
Use -o to set the archive name, or, the base filename
plus the archive extension will be used.

At this stage it is NOT possible to create an archive
with just C++ files: the base Felix file is archived too.

Option checking might not be 100%.
[If there's an exe it run that by mistake:]