Add Env::getenvlist(v) to get lists from environment.
Uses ":" on Unix and ";" on Windows. So we will be able to specify
PKG_CONFIG_PATH as a list instead of just a single directory.
This will allow environment driven "cascading packages" which
in turn allows abstract top level config packages, with
platform specific additions as required.

This should help enable largely readonly configs pre-built
for common platforms without the build system needing to
selectively copy stuff. In turn this parametrisation should
provide initial support for cross compilation by centralising
the platform model into a few discrete components.

Without this every add on package must supply configs
for every available platform, which is just NOT ON.
With this, most packages will only need a generic config.