Clean up the builder a bit more. Add weblink target to Makefile.
The target "weblink" builds a version of the webserver
with pre-loaded plugins, so that it can run as a stand-alone
executable without needing an LD_LIBRRARY_PATH to find plugins.

Some cleanup of naming conventions is required, so that we can
then simplify the specification of the linkage using the
Felix parser. We should then proceed to find some way
to make the specification "more automatic".

Frozen verions of scoop, flx, etc can then be made.
Note at present "flx" does not load any plugins,
however it it is to use the toolchain_* plugins
for compilation to allow cross compiling, it will
need to. And then we need the freezing capability to
"undo" the dynamic loading, at least for the host
system.