Add prototype Felix builder for flxg, the Felix compiler.
At this point several Ocaml libraries build, however
we cheat by copying dypgen instead of building it.
The tool processes ocamllex, ocamlyacc, and dypgen source
files, but with fixed flags suitable for Felix.

Currently bugs out because flx_version_hook is missing
from the repository (it's a generated file). We need to
generate it.

Also in this commit, system() and get_stdout() accept any
Streamable data structure with string value type.
A bug in Filename is fixed (infinite recursion on empty
file name). And we implement Streamable \subseteq Set,
which returns true if all the values of a Streamable
are members of the Set. Hence you can ask if an array
is an improper subset of a list.